Clean up of Deflectometry class. Now works with mostly unqiue Pointers instead of shared ptrs. Also cleaned up a lot of the Image Processing class. Implented some accessor functions to get rid in future of forwarding functions form deflectometry-> to ImageProcessing. Implemented some cleaner Functionts that have only one Purpose. This was necessary. Created GrayCalibVector for Gray Response Curve in Sections and Implmented a GrayCalibration class. For this the Pattern class was stripped from the LUT function. all Gray Value Calib should be in future done by the gray Calibration class. It is hard.

+++ include/camera_calib.hpp
@@ -316,7 +316,7 @@ array<vector<Mat>, 2> runCameraCalibration(vector<Mat>& frames,
     
 
 
-    s.nrFrames = frames.size();
+    s.nrFrames = static_cast<int>(frames.size());
     fs.release();                                         // close Settings file
     //! [file_read]
